+++ crates/teloxide/src/utils/html.rs
@@ -2,7 +2,7 @@
 //!
 //! [spec]: https://core.telegram.org/bots/api#html-style
 
-use teloxide_core::types::User;
+use teloxide_core::types::{User, UserId};
 
 /// Applies the bold font style to the string.
 ///
@@ -56,7 +56,7 @@ pub fn link(url: &str, text: &str) -> String {
 /// Builds an inline user mention link with an anchor.
 #[must_use = "This function returns a new string, rather than mutating the argument, so calling it \
               without using its output does nothing useful"]
-pub fn user_mention(user_id: i64, text: &str) -> String {
+pub fn user_mention(user_id: UserId, text: &str) -> String {
     link(format!("tg://user?id={user_id}").as_str(), text)
 }
 
@@ -158,7 +158,7 @@ mod tests {
     #[test]
     fn test_user_mention() {
         assert_eq!(
-            user_mention(123_456_789, "<pwner666>"),
+            user_mention(UserId(123_456_789), "<pwner666>"),
             "<a href=\"tg://user?id=123456789\">&lt;pwner666&gt;</a>",
         );
     }